requirements.txt
Pillow>=10.0.0
tqdm>=4.64.0
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
wandb>=0.15.0